Move FeatureFlagConnectedInstrumentationProvider to @tloncorp/app
