+++ packages/app/package.json
@@ -15,6 +15,7 @@
     "@types/react-native": "0.73.0"
   },
   "peerDependencies": {
+    "@react-native-firebase/perf": "^19.2.2",
     "lodash": "^4.17.21",
     "zustand": "^3.7.2"
   }
